+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to NetVips will be documented in this file. See [here](CHANG
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [3.2.0] - TBD
+### Added
+- Add generic array overload for `Image.NewFromMemory()`.
+
 ### Changed
 - Remove version constraint for `System.Buffers` when targeting .NET Framework ([#262](https://github.com/kleisauke/net-vips/issues/262)).
 - Lower the minimum version requirements in [NetVips.Extensions](https://www.nuget.org/packages/NetVips.Extensions/) package: